+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Added merging of similar compile requests to the compile queue (#2137)
 - Export all handler's / resource's module's plugin source files so helper functions can be used from sibling modules (#2162)
 - Added documentation on how a string is matched against a regex defined in a regex-based typedef (#2214)
+- Added API to query ResourceActions
 - Added support to query the resource action log of a resource via the CLI (#2253)
 - Added conditional expression to the language with syntax condition ? x: y (#1987)
 - Add support for inmanta-cli click plugins